+++ pyrax/fakes.py
@@ -589,7 +589,7 @@ identity_type = rackspace
 keyring_username =
 region = FAKE
 custom_user_agent = FAKE
-debug =
+http_debug =
 """
 
 # This will handle both singular and plural responses.
@@ -597,159 +597,177 @@ fake_identity_user_response = {
         "users": [{"name": "fake", "id": "fake"},
             {"name": "faker", "id": "faker"}],
         "user": {"name": "fake", "id": "fake"},
-        "roles": [{u'description': u'User Admin Role.',
-                u'id': u'3',
-                u'name': u'identity:user-admin'}],
+        "roles": [{u'description': 'User Admin Role.',
+                'id': '3',
+                'name': 'identity:user-admin'}],
         }
 
 fake_identity_tenant_response = {"name": "fake", "id": "fake",
         "description": "fake", "enabled": True}
 
 fake_identity_tenants_response = {
-        "tenants": [
-                {"name": "fake", "id": "fake", "description": "fake",
-                        "enabled": True},
-                {"name": "faker", "id": "faker", "description": "faker",
-                        "enabled": True},
-                ]}
+    "tenants": [
+        {"name": "fake", "id": "fake", "description": "fake",
+        "enabled": True},
+        {"name": "faker", "id": "faker", "description": "faker",
+        "enabled": True},
+        ]}
 
-fake_identity_tokens_response = {"access": {
-        u'metadata': {u'is_admin': 0,
-          u'roles': [u'asdfgh',
-           u'sdfghj',
-           u'dfghjk']},
-          u'serviceCatalog': [{u'endpoints': [{u'adminURL': u'http://10.0.0.0:8774/v2/qweqweqwe',
-             u'id': u'dddddddddd',
-             u'publicURL': u'http://10.0.0.0:8774/v2/qweqweqwe',
-             u'internalURL': u'http://10.0.0.0:8774/v2/qweqweqwe',
-             u'region': u'some_region'}],
-           u'endpoints_links': [],
-           u'name': u'nova',
-           u'type': u'compute'},
-          {u'endpoints': [{u'adminURL': u'http://10.0.0.0:35357/v2.0',
-             u'id': u'qweqweqwe',
-             u'internalURL': u'http://10.0.0.0:5000/v2.0',
-             u'publicURL': u'http://10.0.0.0:5000/v2.0',
-             u'region': u'some_region'}],
-           u'endpoints_links': [],
-           u'name': u'keystone',
-           u'type': u'identity'}],
-        u'token': {u'expires': u'1999-05-04T16:45:05Z',
-          u'id': u'qweqweqwe',
-          u'tenant': {u'description': u'admin Tenant',
-           u'enabled': True,
-           u'id': u'qweqweqwe',
-           u'name': u'admin'}},
-        u'user': {u'id': u'qweqweqwe',
-          u'name': u'admin',
-          u'roles': [{u'id': u'qweqweqwe', u'name': u'admin'},
-           {u'id': u'qweqweqwe', u'name': u'KeystoneAdmin'},
-           {u'id': u'qweqweqwe',
-            u'name': u'KeystoneServiceAdmin'}],
-          u'roles_links': [],
-          u'username': u'admin'}}}
+fake_identity_tokens_response = {"access":
+        {'metadata': {u'is_admin': 0,
+            'roles': [u'asdfgh',
+                'sdfghj',
+                'dfghjk']},
+        'serviceCatalog': [{u'endpoints': [
+            {u'adminURL': 'http://10.0.0.0:8774/v2/qweqweqwe',
+            'id': 'dddddddddd',
+            'publicURL': 'http://10.0.0.0:8774/v2/qweqweqwe',
+            'internalURL': 'http://10.0.0.0:8774/v2/qweqweqwe',
+            'region': 'some_region'}],
+            'endpoints_links': [],
+            'name': 'nova',
+            'type': 'compute'},
+            {u'endpoints': [{u'adminURL': 'http://10.0.0.0:35357/v2.0',
+            'id': 'qweqweqwe',
+            'internalURL': 'http://10.0.0.0:5000/v2.0',
+            'publicURL': 'http://10.0.0.0:5000/v2.0',
+            'region': 'some_region'}],
+            'endpoints_links': [],
+            'name': 'keystone',
+            'type': 'identity'}],
+        'token': {u'expires': '1999-05-04T16:45:05Z',
+            'id': 'qweqweqwe',
+            'tenant': {u'description': 'admin Tenant',
+                'enabled': True,
+                'id': 'qweqweqwe',
+                'name': 'admin'}},
+        'user': {u'id': 'qweqweqwe',
+            'name': 'admin',
+            'roles': [{u'id': 'qweqweqwe', 'name': 'admin'},
+                {u'id': 'qweqweqwe', 'name': 'KeystoneAdmin'},
+                {u'id': 'qweqweqwe',
+                'name': 'KeystoneServiceAdmin'}],
+            'roles_links': [],
+            'username': 'admin'}}}
 
 fake_identity_endpoints_response = {"access": {
         "endpoints": ["fake", "faker", "fakest"]}}
 
 fake_identity_response = {u'access':
-        {u'serviceCatalog': [
-            {u'endpoints': [{u'publicURL': u'https://ord.loadbalancers.api.rackspacecloud.com/v1.0/000000',
-                             u'region': u'ORD',
-                             u'tenantId': u'000000'},
-                            {u'publicURL': u'https://dfw.loadbalancers.api.rackspacecloud.com/v1.0/000000',
-                             u'region': u'DFW',
-                             u'tenantId': u'000000'},
-                             {u'publicURL': u'https://syd.loadbalancers.api.rackspacecloud.com/v1.0/000000',
-                             u'region': u'SYD',
-                             u'tenantId': u'000000'}],
-             u'name': u'cloudLoadBalancers',
-             u'type': u'rax:load-balancer'},
-            {u'endpoints': [{u'internalURL': u'https://snet-storage101.fake1.clouddrive.com/v1/MossoCloudFS_ffffffff-ffff-ffff-ffff-ffffffffffff',
-                             u'publicURL': u'https://storage101.fake1.clouddrive.com/v1/MossoCloudFS_ffffffff-ffff-ffff-ffff-ffffffffffff',
-                             u'region': u'FAKE',
-                             u'tenantId': u'MossoCloudFS_ffffffff-ffff-ffff-ffff-ffffffffffff'},
-                            {u'internalURL': u'https://snet-storage101.dfw1.clouddrive.com/v1/MossoCloudFS_ffffffff-ffff-ffff-ffff-ffffffffffff',
-                             u'publicURL': u'https://storage101.dfw1.clouddrive.com/v1/MossoCloudFS_ffffffff-ffff-ffff-ffff-ffffffffffff',
-                             u'region': u'DFW',
-                             u'tenantId': u'MossoCloudFS_ffffffff-ffff-ffff-ffff-ffffffffffff'},
-                            {u'internalURL': u'https://snet-storage101.ord1.clouddrive.com/v1/MossoCloudFS_ffffffff-ffff-ffff-ffff-ffffffffffff',
-                             u'publicURL': u'https://storage101.ord1.clouddrive.com/v1/MossoCloudFS_ffffffff-ffff-ffff-ffff-ffffffffffff',
-                             u'region': u'ORD',
-                             u'tenantId': u'MossoCloudFS_ffffffff-ffff-ffff-ffff-ffffffffffff'},
-                             {u'internalURL': u'https://snet-storage101.syd1.clouddrive.com/v1/MossoCloudFS_ffffffff-ffff-ffff-ffff-ffffffffffff',
-                             u'publicURL': u'https://storage101.ord1.clouddrive.com/v1/MossoCloudFS_ffffffff-ffff-ffff-ffff-ffffffffffff',
-                             u'region': u'SYD',
-                             u'tenantId': u'MossoCloudFS_ffffffff-ffff-ffff-ffff-ffffffffffff'}],
-             u'name': u'cloudFiles',
-             u'type': u'object-store'},
-            {u'endpoints': [{u'publicURL': u'https://dfw.servers.api.rackspacecloud.com/v2/000000',
-                             u'region': u'DFW',
-                             u'tenantId': u'000000',
-                             u'versionId': u'2',
-                             u'versionInfo': u'https://dfw.servers.api.rackspacecloud.com/v2',
-                             u'versionList': u'https://dfw.servers.api.rackspacecloud.com/'},
-                            {u'publicURL': u'https://ord.servers.api.rackspacecloud.com/v2/000000',
-                             u'region': u'ORD',
-                             u'tenantId': u'000000',
-                             u'versionId': u'2',
-                             u'versionInfo': u'https://ord.servers.api.rackspacecloud.com/v2',
-                             u'versionList': u'https://ord.servers.api.rackspacecloud.com/'},
-                             {u'publicURL': u'https://syd.servers.api.rackspacecloud.com/v2/000000',
-                             u'region': u'SYD',
-                             u'tenantId': u'000000',
-                             u'versionId': u'2',
-                             u'versionInfo': u'https://syd.servers.api.rackspacecloud.com/v2',
-                             u'versionList': u'https://syd.servers.api.rackspacecloud.com/'}],
-             u'name': u'cloudServersOpenStack',
-             u'type': u'compute'},
-            {u'endpoints': [{u'publicURL': u'https://dns.api.rackspacecloud.com/v1.0/000000',
-                             u'tenantId': u'000000'}],
-             u'name': u'cloudDNS',
-             u'type': u'rax:dns'},
-            {u'endpoints': [{u'publicURL': u'https://dfw.databases.api.rackspacecloud.com/v1.0/000000',
-                             u'region': u'DFW',
-                             u'tenantId': u'000000'},
-                             {u'publicURL': u'https://syd.databases.api.rackspacecloud.com/v1.0/000000',
-                             u'region': u'SYD',
-                             u'tenantId': u'000000'},
-                            {u'publicURL': u'https://ord.databases.api.rackspacecloud.com/v1.0/000000',
-                             u'region': u'ORD',
-                             u'tenantId': u'000000'}],
-             u'name': u'cloudDatabases',
-             u'type': u'rax:database'},
-            {u'endpoints': [{u'publicURL': u'https://servers.api.rackspacecloud.com/v1.0/000000',
-                             u'tenantId': u'000000',
-                             u'versionId': u'1.0',
-                             u'versionInfo': u'https://servers.api.rackspacecloud.com/v1.0',
-                             u'versionList': u'https://servers.api.rackspacecloud.com/'}],
-             u'name': u'cloudServers',
-             u'type': u'compute'},
-            {u'endpoints': [{u'publicURL': u'https://cdn1.clouddrive.com/v1/MossoCloudFS_ffffffff-ffff-ffff-ffff-ffffffffffff',
-                             u'region': u'DFW',
-                             u'tenantId': u'MossoCloudFS_ffffffff-ffff-ffff-ffff-ffffffffffff'},
-                             {u'publicURL': u'https://cdn1.clouddrive.com/v1/MossoCloudFS_ffffffff-ffff-ffff-ffff-ffffffffffff',
-                             u'region': u'SYD',
-                             u'tenantId': u'MossoCloudFS_ffffffff-ffff-ffff-ffff-ffffffffffff'},
-                            {u'publicURL': u'https://cdn2.clouddrive.com/v1/MossoCloudFS_ffffffff-ffff-ffff-ffff-ffffffffffff',
-                             u'region': u'ORD',
-                             u'tenantId': u'MossoCloudFS_ffffffff-ffff-ffff-ffff-ffffffffffff'}],
-             u'name': u'cloudFilesCDN',
-             u'type': u'rax:object-cdn'},
-            {u'endpoints': [{u'publicURL': u'https://monitoring.api.rackspacecloud.com/v1.0/000000',
-                             u'tenantId': u'000000'}],
-             u'name': u'cloudMonitoring',
-             u'type': u'rax:monitor'}],
-u'token': {u'expires': u'2222-02-22T22:22:22.000-02:00',
-    u'id': u'xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx',
-    u'tenant': {u'id': u'000000', u'name': u'000000'}},
-u'user': {u'id': u'123456',
-    u'name': u'fakeuser',
-    u'RAX-AUTH:defaultRegion': u'DFW',
-    u'roles': [{u'description': u'User Admin Role.',
-            u'id': u'3',
-            u'name': u'identity:user-admin'}],
-            }}}
+    {u'serviceCatalog': [
+        {u'endpoints': [{u'publicURL':
+            'https://ord.loadbalancers.api.rackspacecloud.com/v1.0/000000',
+            'region': 'ORD',
+            'tenantId': '000000'},
+            {u'publicURL':
+            'https://dfw.loadbalancers.api.rackspacecloud.com/v1.0/000000',
+            'region': 'DFW',
+            'tenantId': '000000'},
+            {u'publicURL':
+            'https://syd.loadbalancers.api.rackspacecloud.com/v1.0/000000',
+            'region': 'SYD',
+            'tenantId': '000000'}],
+        'name': 'cloudLoadBalancers',
+        'type': 'rax:load-balancer'},
+        {u'endpoints': [{u'internalURL':
+            'https://snet-aa.fake1.clouddrive.com/v1/MossoCloudFS_abc',
+            'publicURL': 'https://aa.fake1.clouddrive.com/v1/MossoCloudFS_abc',
+            'region': 'FAKE',
+            'tenantId': 'MossoCloudFS_abc'},
+            {u'internalURL':
+            'https://snet-aa.dfw1.clouddrive.com/v1/MossoCloudFS_abc',
+            'publicURL': 'https://aa.dfw1.clouddrive.com/v1/MossoCloudFS_abc',
+            'region': 'DFW',
+            'tenantId': 'MossoCloudFS_abc'},
+            {u'internalURL':
+            'https://snet-aa.ord1.clouddrive.com/v1/MossoCloudFS_abc',
+            'publicURL': 'https://aa.ord1.clouddrive.com/v1/MossoCloudFS_abc',
+            'region': 'ORD',
+            'tenantId': 'MossoCloudFS_abc'},
+            {u'internalURL':
+            'https://snet-aa.syd1.clouddrive.com/v1/MossoCloudFS_abc',
+            'publicURL': 'https://aa.ord1.clouddrive.com/v1/MossoCloudFS_abc',
+            'region': 'SYD',
+            'tenantId': 'MossoCloudFS_abc'}],
+        'name': 'cloudFiles',
+        'type': 'object-store'},
+        {u'endpoints': [{u'publicURL':
+            'https://dfw.servers.api.rackspacecloud.com/v2/000000',
+            'region': 'DFW',
+            'tenantId': '000000',
+            'versionId': '2',
+            'versionInfo': 'https://dfw.servers.api.rackspacecloud.com/v2',
+            'versionList': 'https://dfw.servers.api.rackspacecloud.com/'},
+            {u'publicURL':
+            'https://ord.servers.api.rackspacecloud.com/v2/000000',
+            'region': 'ORD',
+            'tenantId': '000000',
+            'versionId': '2',
+            'versionInfo': 'https://ord.servers.api.rackspacecloud.com/v2',
+            'versionList': 'https://ord.servers.api.rackspacecloud.com/'},
+            {u'publicURL':
+            'https://syd.servers.api.rackspacecloud.com/v2/000000',
+            'region': 'SYD',
+            'tenantId': '000000',
+            'versionId': '2',
+            'versionInfo': 'https://syd.servers.api.rackspacecloud.com/v2',
+            'versionList': 'https://syd.servers.api.rackspacecloud.com/'}],
+        'name': 'cloudServersOpenStack',
+        'type': 'compute'},
+        {u'endpoints': [{u'publicURL':
+        'https://dns.api.rackspacecloud.com/v1.0/000000',
+        'tenantId': '000000'}],
+        'name': 'cloudDNS',
+        'type': 'rax:dns'},
+        {u'endpoints': [{u'publicURL':
+            'https://dfw.databases.api.rackspacecloud.com/v1.0/000000',
+            'region': 'DFW',
+            'tenantId': '000000'},
+            {u'publicURL':
+            'https://syd.databases.api.rackspacecloud.com/v1.0/000000',
+            'region': 'SYD',
+            'tenantId': '000000'},
+            {u'publicURL':
+            'https://ord.databases.api.rackspacecloud.com/v1.0/000000',
+            'region': 'ORD',
+            'tenantId': '000000'}],
+        'name': 'cloudDatabases',
+        'type': 'rax:database'},
+        {u'endpoints': [{u'publicURL':
+        'https://servers.api.rackspacecloud.com/v1.0/000000',
+        'tenantId': '000000',
+        'versionId': '1.0',
+        'versionInfo': 'https://servers.api.rackspacecloud.com/v1.0',
+        'versionList': 'https://servers.api.rackspacecloud.com/'}],
+        'name': 'cloudServers',
+        'type': 'compute'},
+        {u'endpoints': [{u'publicURL':
+            'https://cdn1.clouddrive.com/v1/MossoCloudFS_abc',
+            'region': 'DFW',
+            'tenantId': 'MossoCloudFS_abc'},
+            {u'publicURL': 'https://cdn1.clouddrive.com/v1/MossoCloudFS_abc',
+            'region': 'SYD',
+            'tenantId': 'MossoCloudFS_abc'},
+            {u'publicURL': 'https://cdn2.clouddrive.com/v1/MossoCloudFS_abc',
+            'region': 'ORD',
+            'tenantId': 'MossoCloudFS_abc'}],
+        'name': 'cloudFilesCDN',
+        'type': 'rax:object-cdn'},
+        {u'endpoints': [{u'publicURL':
+            'https://monitoring.api.rackspacecloud.com/v1.0/000000',
+            'tenantId': '000000'}],
+        'name': 'cloudMonitoring',
+        'type': 'rax:monitor'}],
+    u'token': {u'expires': '2222-02-22T22:22:22.000-02:00',
+    'id': 'xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx',
+    'tenant': {u'id': '000000', 'name': '000000'}},
+    u'user': {u'id': '123456',
+    'name': 'fakeuser',
+    'RAX-AUTH:defaultRegion': 'DFW',
+    'roles': [{u'description': 'User Admin Role.',
+    'id': '3',
+    'name': 'identity:user-admin'}],
+    }}}
 
 
 
@@ -774,8 +792,12 @@ class FakeIdentityResponse(FakeResponse):
     def read(self):
         return json.dumps(self.content)
 
-_module_pth = os.path.dirname(pyrax.__file__)
-_img_path = os.path.join(_module_pth, "..", "tests", "unit", "python-logo.png")
-png_content = None
-with open(_img_path, "rb") as pfile:
-    png_content = pfile.read()
+
+def get_png_content():
+    _module_pth = os.path.dirname(pyrax.__file__)
+    _img_path = os.path.join(_module_pth, "..", "tests", "unit",
+            "python-logo.png")
+    png_content = None
+    with open(_img_path, "rb") as pfile:
+        png_content = pfile.read()
+    return png_content